+++ src/test/java/com/mageddo/dnsproxyserver/server/dns/solver/docker/application/ContainerSolvingServiceTest.java
@@ -1,42 +1,48 @@
-// fixme
-//package com.mageddo.dnsproxyserver.server.dns.solver.docker.application;
-//
-//import com.mageddo.dnsproxyserver.docker.DockerFacade;
-//import com.mageddo.dnsproxyserver.docker.DockerNetworkFacade;
-//import com.mageddo.net.IP;
-//import org.junit.jupiter.api.Test;
-//import org.junit.jupiter.api.extension.ExtendWith;
-//import org.mockito.InjectMocks;
-//import org.mockito.Mock;
-//import org.mockito.Spy;
-//import org.mockito.junit.jupiter.MockitoExtension;
-//import testing.templates.IpTemplates;
-//
-//import static org.junit.jupiter.api.Assertions.assertEquals;
-//import static org.junit.jupiter.api.Assertions.assertNotNull;
-//import static org.junit.jupiter.api.Assertions.assertNull;
-//import static org.mockito.ArgumentMatchers.eq;
-//import static org.mockito.Mockito.doReturn;
-//import static org.mockito.Mockito.never;
-//import static org.mockito.Mockito.verify;
-//import static testing.templates.docker.InspectContainerResponseTemplates.ngixWithDefaultBridgeNetworkOnly;
-//
-//// todo #444
-//@ExtendWith(MockitoExtension.class)
-//class ContainerSolvingServiceTest {
-//
-//  @Mock
-//  DockerFacade dockerFacade;
-//
-//  @Mock
-//  DockerNetworkFacade networkDAO;
-//
-//  @Mock
-//  MatchingContainerService matchingContainerService;
-//
-//  @Spy
-//  @InjectMocks
-//  ContainerSolvingServiceTest containerSolvingService;
+package com.mageddo.dnsproxyserver.server.dns.solver.docker.application;
+
+import com.mageddo.dnsproxyserver.docker.DockerFacade;
+import com.mageddo.dnsproxyserver.docker.DockerNetworkFacade;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Spy;
+import org.mockito.junit.jupiter.MockitoExtension;
+import testing.templates.server.dns.solver.docker.ContainerTemplates;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+// todo #444
+@ExtendWith(MockitoExtension.class)
+class ContainerSolvingServiceTest {
+
+  @Mock
+  DockerFacade dockerFacade;
+
+  @Mock
+  DockerNetworkFacade networkDAO;
+
+  @Mock
+  MatchingContainerService matchingContainerService;
+
+  @Spy
+  @InjectMocks
+  ContainerSolvingService containerSolvingService;
+
+  @Test
+  void mustSolveSpecifiedNetworkFirst() {
+    // arrange
+    final var container = ContainerTemplates.withDpsLabel();
+
+    // act
+    final var ip = this.containerSolvingService.findBestIpMatch(container);
+
+    // assert
+    assertNotNull(ip);
+    assertEquals("172.23.0.2", ip);
+
+  }
 //
 //  @Test
 //  void mustReturnHostMachineIPWhenThereIsNoBetterMatch() {
@@ -85,5 +91,5 @@
 //    verify(this.dockerFacade, never()).findHostMachineIp(eq(version));
 //
 //  }
-//
-//}
+
+}
